Add physical device inspection C++ code
