+++ .clang-format
@@ -14,8 +14,10 @@ AlignTrailingComments: true
 AllowShortBlocksOnASingleLine: Empty
 AllowShortCaseLabelsOnASingleLine: false
 AllowShortFunctionsOnASingleLine: Empty
+AllowShortIfStatementsOnASingleLine: Never
 AlwaysBreakTemplateDeclarations: Yes
 IndentPPDirectives: BeforeHash
+PPIndentWidth: 4
 AlignAfterOpenBracket: BlockIndent
 AllowAllArgumentsOnNextLine: true
 AllowAllParametersOfDeclarationOnNextLine: true
@@ -23,3 +25,14 @@ BinPackParameters: false
 BinPackArguments: false
 EmptyLineAfterAccessModifier: Never
 EmptyLineBeforeAccessModifier: Always
+AlignEscapedNewlines: Left
+AlignOperands: AlignAfterOperator
+AllowShortEnumsOnASingleLine: false
+AllowShortLambdasOnASingleLine: Empty
+BreakBeforeBraces: Attach
+BreakConstructorInitializers: AfterColon
+BreakStringLiterals: true
+MaxEmptyLinesToKeep: 1
+QualifierAlignment: Left
+SeparateDefinitionBlocks: Always
+UseTab: Never
